get_callinfo_and: don't search by memory address
The return value of malloc(), even with ASLR disabled, is heavily
dependent on the state of the program, which is affected by object
tracking. For example, when using LD_PRELOAD=libobjtracker.so <prog>
to track all objects, the allocations made by the object tracker
will be different when we run it again to only track the calls to
malloc() listed in <prog.objtrack>. Therefore, disable checks
against the return address for now until we can figure out a way to
make object tracking affect malloc() the same way, regardless of
configuration.