Implement oracle; fix bus error.
Fixed bus error on platforms that don't support concurrent UM access.
See
https://devtalk.nvidia.com/default/topic/1031749/jetson-tx1/concurrent-cpu-and-gpu-processing/
and
https://docs.nvidia.com/cuda/cublas/index.html#cublas-level-1-function-reference
for more information.

The issue may not be completely fixed as it's possible for UM buffers to
be accessed outside of the Object Tracker code on a separate thread
while a kernel call is in progress.

Added "oracle" decision maker, which uses a trace to determine which
malloc()s to substitute for a UM buffer.